Change LogEntry TimeSpent to float

+++ ActivityTracker/Controllers/ActivitiesController.cs
@@ -154,7 +154,7 @@ namespace ActivityTracker.Controllers
                 OrderByDescending(le => le.Date).ToList();
             var logEntries = (IEnumerable<dynamic>)ViewBag.LogEntries;
             ViewBag.StartDate = logEntries.Min(x => x.Date);
-            ViewBag.TimeSpent = logEntries.Sum(x => x.TimeSpent);
+            ViewBag.TimeSpent = logEntries.Sum(x => (float)x.TimeSpent);
             ViewBag.LastActivity = logEntries.Max(x => x.Date);
 
             var activity = await _context.Activities.SingleOrDefaultAsync(m => m.ID == id);
